scanner: Fix build errors
- Fix getCenterFrequency method by using waterfall's center frequency
- Fix string conversion issue with size_t to std::string

+++ misc_modules/scanner/src/ScannerPSD.cpp
@@ -170,7 +170,7 @@ bool ScannerPSD::process() {
 bool ScannerPSD::processFrame(const std::vector<std::complex<float>>& frame) {
     if (!m_initialized || frame.size() < m_fftSize) {
         flog::error("Scanner: Cannot process frame - initialized: {}, frame size: {} (need {})",
-                           m_initialized ? "true" : "false", frame.size(), m_fftSize);
+                           m_initialized ? "true" : "false", static_cast<int>(frame.size()), m_fftSize);
         return false;
     }
